ml/requirements_ml.txt
tensorflow==2.20.0
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.1.78
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
